initial attempt to refactoring _HDFFile class
