--- src/ophyd_async/epics/areadetector/writers/_hdfdataset.py
@@ -1,10 +1,0 @@
-from dataclasses import dataclass
-from typing import Sequence
-
-
-@dataclass
-class _HDFDataset:
-    name: str
-    path: str
-    shape: Sequence[int]
-    multiplier: int
