+++ tests/panda/test_writer.py
@@ -15,6 +15,10 @@ from ophyd_async.epics.pvi import create_children_from_annotations, fill_pvi_ent
 from ophyd_async.epics.signal.signal import epics_signal_r
 from ophyd_async.panda import CommonPandaBlocks
 from ophyd_async.panda.writers._hdf_writer import (
+from ophyd_async.epics.areadetector.writers.general_hdffile import _HDFFile
+
+from ophyd_async.panda.writers import PandaHDFWriter
+from ophyd_async.panda.writers.hdf_writer import (
     Capture,
     CaptureSignalWrapper,
     PandaHDFWriter,
